requirements.txt
fastapi
uvicorn
openai
yt-dlp
ffmpeg-python
python-multipart
openai-whisper 
python-dotenv
